pyproject.toml
[project]
name = "cql-agent"
version = "0.1.0"
description = "AI Agent for Home Repairs"
requires-python = ">=3.12,<3.13"
dependencies = [
    "spacy>=3.7.0",
    "pt_core_news_sm @ https://github.com/explosion/spacy-models/releases/download/pt_core_news_sm-3.7.0/pt_core_news_sm-3.7.0-py3-none-any.whl",
    "numpy>=1.26.0,<2.0.0",
    "chromadb>=1.1.1",
    "duckduckgo-search>=8.1.1",
    "langchain>=0.3.27",
    "langchain-chroma>=0.2.6",
    "langchain-community>=0.3.31",
    "langchain-ollama>=0.3.10",
    "langgraph>=0.2.0",
    "pydantic>=2.12.2",
    "pypdf>=6.1.3",
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.32.0",
    "python-multipart>=0.0.9",
    "requests>=2.32.0",
    "bleach>=6.2.0",
    "markupsafe>=3.0.0",
    "sqlparse>=0.5.0",
    "pytest>=8.0.0",
    "pytest-cov>=4.1.0",
    "pytest-asyncio>=0.21.0",
    "httpx>=0.27.0",
    "urllib3>=2.5.0",
    "flake8>=7.0.0",
    "pylint>=3.0.0",
    "rapidfuzz>=3.14.1",
    "pip-audit>=2.6.0",
    "safety>=3.0.0",
    "python-json-logger>=3.2.1",
    "redis>=5.0.0",
    "python-dotenv>=1.1.1",
    "pyjwt>=2.8.0",
]

[project.optional-dependencies]
openai = [
    "langchain-openai>=0.2.14",
]
google = [
    "langchain-google-genai>=2.0.8",
]
anthropic = [
    "langchain-anthropic>=0.3.5",
]
all-providers = [
    "langchain-openai>=0.2.14",
    "langchain-google-genai>=2.0.8",
    "langchain-anthropic>=0.3.5",
]

[tool.pytest.ini_options]
testpaths = ["api/tests"]
python_files = ["test_*.py", "*_test.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = [
    "-v",
    "--strict-markers",
    "--tb=short",
    "--cov=api",
    "--cov=agents",
    "--cov-report=html",
    "--cov-report=term-missing",
]
markers = [
    "slow: marks tests as slow",
    "integration: marks tests as integration tests",
    "unit: marks tests as unit tests",
]
filterwarnings = [
    "ignore::DeprecationWarning:spacy.*:",
    "ignore::DeprecationWarning:weasel.*:",
]
